refactor: unify training utilities and simplify cifar loop
